feat: allow decimal carb percentage entry

+++ app/src/main/java/com/medavox/diabeticdiary/carbcalculator/CarbIngredient.kt
@@ -4,10 +4,10 @@ package com.medavox.diabeticdiary.carbcalculator
  * @author Adam Howard
  * @since 01/09/2017
  */
-data class CarbIngredient(val grams:Int, val percentCarb:Int, val name:String?=null) {
+data class CarbIngredient(val grams:Int, val percentCarb:Float, val name:String?=null) {
 
     /**Calculate and return the total CP in this ingredient*/
-    val CPx1000:Int get() {
+    val CPx1000:Float get() {
         return percentCarb * grams
     }
